+++ pkg/plugin/client_query_test.go
@@ -7,50 +7,51 @@ import (
 	"time"
 
 	"github.com/d-velop/grafana-odata-datasource/pkg/plugin/odata"
-	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestMapFilter(t *testing.T) {
 	tables := []struct {
 		name             string
-		timeProperty     string
-		timeRange        backend.TimeRange
 		filterConditions []filterCondition
 		expected         string
 	}{
 		{
-			name:             "Time filter only",
-			timeProperty:     "time",
-			timeRange:        aOneDayTimeRange(),
-			filterConditions: someFilterConditions(int32Eq5),
-			expected:         "time ge 2022-04-21T12:30:50Z and time le 2022-04-21T12:30:50Z and int32 eq 5",
+			name: "Time filter only",
+			filterConditions: someFilterConditions(
+				withFilterCondition(timeProp, "ge", aOneDayTimeRange().From.Format(time.RFC3339)),
+				withFilterCondition(timeProp, "le", aOneDayTimeRange().To.Format(time.RFC3339)),
+				int32Eq5),
+			expected: "time ge 2022-04-21T12:30:50Z and time le 2022-04-21T12:30:50Z and int32 eq 5",
 		},
 		{
-			name:             "Time filter and int and string filter",
-			timeProperty:     "time",
-			timeRange:        aOneDayTimeRange(),
-			filterConditions: someFilterConditions(int32Eq5, withFilterCondition(stringProp, "eq", "Hello")),
-			expected:         "time ge 2022-04-21T12:30:50Z and time le 2022-04-21T12:30:50Z and int32 eq 5 and string eq 'Hello'",
+			name: "Time filter and int and string filter",
+			filterConditions: someFilterConditions(
+				withFilterCondition(timeProp, "ge", aOneDayTimeRange().From.Format(time.RFC3339)),
+				withFilterCondition(timeProp, "le", aOneDayTimeRange().To.Format(time.RFC3339)),
+				int32Eq5,
+				withFilterCondition(stringProp, "eq", "Hello")),
+			expected: "time ge 2022-04-21T12:30:50Z and time le 2022-04-21T12:30:50Z and int32 eq 5 and string eq 'Hello'",
 		},
 		{
-			name:             "Time filter and string filter",
-			timeProperty:     "time",
-			timeRange:        aOneDayTimeRange(),
-			filterConditions: someFilterConditions(withFilterCondition(stringProp, "eq", "")),
-			expected:         "time ge 2022-04-21T12:30:50Z and time le 2022-04-21T12:30:50Z and string eq ''",
+			name: "Time filter and string filter",
+			filterConditions: someFilterConditions(
+				withFilterCondition(timeProp, "ge", aOneDayTimeRange().From.Format(time.RFC3339)),
+				withFilterCondition(timeProp, "le", aOneDayTimeRange().To.Format(time.RFC3339)),
+				withFilterCondition(stringProp, "eq", "")),
+			expected: "time ge 2022-04-21T12:30:50Z and time le 2022-04-21T12:30:50Z and string eq ''",
 		},
 		{
 			name:             "String filter only",
 			filterConditions: someFilterConditions(withFilterCondition(stringProp, "eq", "")),
-			expected:         " and string eq ''",
+			expected:         "string eq ''",
 		},
 	}
 
 	for _, table := range tables {
 		t.Run(table.name, func(t *testing.T) {
 			// Act
-			var filterString = mapFilter(table.timeProperty, table.timeRange, table.filterConditions)
+			var filterString = mapFilter(table.filterConditions)
 
 			// Assert
 			assert.Equal(t, table.expected, filterString)
@@ -65,27 +66,27 @@ func TestBuildQueryUrl(t *testing.T) {
 		entitySet        string
 		properties       []property
 		timeProperty     string
-		timeRange        backend.TimeRange
+		timeRange        []filterCondition
 		filterConditions []filterCondition
 		expected         string
 	}{
 		{
-			name:             "Success",
-			baseUrl:          "http://localhost:5000",
-			entitySet:        "Temperatures",
-			properties:       []property{aProperty(int32Prop)},
-			timeProperty:     "time",
-			timeRange:        aOneDayTimeRange(),
-			filterConditions: someFilterConditions(withFilterCondition(stringProp, "eq", "")),
-			expected:         "http://localhost:5000/Temperatures?%24filter=time+ge+2022-04-21T12%3A30%3A50Z+and+time+le+2022-04-21T12%3A30%3A50Z+and+string+eq+%27%27&%24select=int32%2Ctime",
+			name:       "Success",
+			baseUrl:    "http://localhost:5000",
+			entitySet:  "Temperatures",
+			properties: []property{aProperty(int32Prop), aProperty(timeProp)},
+			filterConditions: someFilterConditions(
+				withFilterCondition(timeProp, "ge", aOneDayTimeRange().From.Format(time.RFC3339)),
+				withFilterCondition(timeProp, "le", aOneDayTimeRange().To.Format(time.RFC3339)),
+				withFilterCondition(stringProp, "eq", "")),
+			expected: "http://localhost:5000/Temperatures?%24filter=time+ge+2022-04-21T12%3A30%3A50Z+and+time+le+2022-04-21T12%3A30%3A50Z+and+string+eq+%27%27&%24select=int32%2Ctime",
 		},
 	}
 
 	for _, table := range tables {
 		t.Run(table.name, func(t *testing.T) {
 			// Act
-			var builtUrl, err = buildQueryUrl(table.baseUrl, table.entitySet, table.properties, table.timeProperty,
-				table.timeRange, table.filterConditions, "+")
+			var builtUrl, err = buildQueryUrl(table.baseUrl, table.entitySet, table.properties, table.filterConditions, "+")
 
 			// Assert
 			assert.NoError(t, err)
@@ -131,7 +132,7 @@ func TestGetEntities(t *testing.T) {
 			client := GetOC("*", table.handlerCallback)
 
 			// Act
-			resp, err := client.Get("Temperatures", []property{aProperty(int32Prop)}, "time", aOneDayTimeRange(), someFilterConditions(int32Eq5))
+			resp, err := client.Get("Temperatures", []property{aProperty(int32Prop)}, someFilterConditions(int32Eq5))
 
 			// Assert
 			if table.expectedError == nil {
@@ -160,7 +161,7 @@ func TestGetMetadata(t *testing.T) {
 			expectedRespCode: 200,
 			handlerCallback: func(w http.ResponseWriter, r *http.Request) {
 				w.WriteHeader(http.StatusOK)
-				w.Write([]byte("<?xml version=\"1.0\" encoding=\"utf-8\"?><edmx:Edmx Version=\"4.0\" xmlns:edmx=\"http://docs.oasis-open.org/odata/ns/edmx\"></edmx:Edmx>"))
+				w.Write([]byte("<?xml version=\"1.0\" encoding=\"utf-8\"?><edmx:Edmx Version=\"4.0\" xmlns:edmx=\"https://docs.oasis-open.org/odata/ns/edmx\"></edmx:Edmx>"))
 			},
 		},
 		{