+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Change Log
 
+## [1.1.0] 2024-03-14
+
+### Features
+- [#2](https://github.com/d-velop/grafana-odata-datasource/issues/2): Made time property optional
+- Updated to go version `1.21`
+- Updated to Grafana `10.2`
+- Configurable space encoding (`%20`or `+`)
+
 ## [1.0.0] 2023-05-12
+
+### Features
 - Initial revision